refactor(process)：迁移sys_get_rusage系统调用
